requirements.txt
requests==2.25.1
web.py==0.62
pyserial~=3.5